+++ Package.swift
@@ -4,7 +4,7 @@ import PackageDescription
 let package = Package(
     name: "Islet",
     platforms: [
-        .macOS("26.0")
+        .macOS("13.0")
     ],
     products: [
         .executable(name: "Islet", targets: ["Islet"])